+++ R/AAA.R
@@ -4,13 +4,14 @@
 #' @importFrom purrr flatten flatten_chr imap
 #' @importFrom withr with_environment
 #' @importFrom assertthat assert_that
-#' @importFrom ggplot2 geom_line geom_point ggplot aes_string xlab ylab
+#' @importFrom ggplot2 geom_line geom_point ggplot xlab ylab
 #' @importFrom ggplot2 guides theme_bw scale_y_continuous scale_y_log10
 #' @importFrom ggplot2 facet_grid facet_wrap aes scale_color_discrete theme
-#' @importFrom ggplot2 scale_color_viridis_c scale_color_brewer
+#' @importFrom ggplot2 scale_color_viridis_c scale_color_brewer labs
+#' @importFrom ggplot2 scale_color_manual
 #' @importFrom patchwork wrap_plots
 #' @importFrom rlang quos sym set_names enexpr := as_string .env abort warn
-#' @importFrom rlang abort is_integerish is_named 
+#' @importFrom rlang abort is_integerish is_named
 #' @importFrom tidyr unnest nest pivot_longer
 #' @importFrom tibble tibble as_tibble
 #' @importMethodsFrom mrgsolve as.list param update as.data.frame
@@ -18,35 +19,41 @@
 #' @importFrom stats as.formula
 #' @importFrom graphics plot
 #' @importFrom glue glue
+#' @importFrom lattice trellis.par.get
+#' @importFrom grDevices hcl.colors
 #' @import methods
-#' 
+#'
 #' @include utils.R
 #' @include sens.R
 #' @include parseq.R
-#' 
-#' 
+#'
+#'
 NULL
 
 #' Sensitivity Analysis with 'mrgsolve'
-#' 
-#' Perform local sensitivity analysis on ordinary differential 
-#' equation based models, including ad-hoc graphical analyses based on 
-#' sequences of parameters as well as local sensitivity analysis. Functions 
+#'
+#' Perform local sensitivity analysis on ordinary differential
+#' equation based models, including ad-hoc graphical analyses based on
+#' sequences of parameters as well as local sensitivity analysis. Functions
 #' are provided for creating inputs, simulating scenarios and plotting outputs.
-#' 
+#'
 #' @details
-#' 
-#' - Local sensitivity analysis: [lsa()]
+#'
+#' - Local sensitivity analysis: [lsa()], [lsa_plot()]
 #' - Run ad-hoc sensitivity analyses: [sens_each()], [sens_grid()], [sens_run()]
 #'   - Use [sens_each_data()] and [sens_grid_data()] to pass in data sets
-#' - Parameter sequence generation: 
+#' - Parameter sequence generation:
 #'   - In a pipeline: [parseq_cv()], [parseq_fct()], [parseq_range()], [parseq_manual()]
 #'   - Stand alone: [seq_cv()], [seq_fct()], [seq_geo()], [seq_even()]
 #' - Plot ad-hoc sensitivity analysis results
 #'   - Use [sens_plot()]
-#' - Select a subset of sensitivity analysis results
-#'   - Use [select_sens()]
-#' 
+#' - Select parameters or results
+#'   - Use [select_par()] to choose which parameters to vary
+#'   - Use [select_sens()] to subset sensitivity analysis results
+#'
+#' @seealso
+#' `vignette("mrgsim.sa", package = "mrgsim.sa")` for a complete tutorial.
+#'
 #' @rdname mrgsim.sa
 #' @name mrgsim.sa
 NULL