+++ R/sens_plot.R
@@ -20,6 +20,64 @@ sens_color_n <- function(data, group) {
   data
 }
 
+pick_palette <- function(ncol, name = ggplot2::waiver()) {
+  colors <- trellis.par.get("superpose.symbol")$col
+  ntrel <- length(colors)
+  if(ncol <= ntrel) {
+    palette <- scale_color_manual(values = colors, name = name)
+    return(palette)
+  } else {
+    colors <- hcl.colors(ncol, "Dark 2")
+    palette <- scale_color_manual(values = colors, name = name)
+  }
+  palette
+}
+
+sens_grid_plot_vars <- function(pars, group = NULL, facet = NULL) {
+  checkvar <- function(x, pars) {
+    if(!is.element(x, pars)) {
+      msg <- glue("`{x}` is not a sensitivity parameter.")
+      abort(msg)  
+    }
+  }
+  if(is.null(group)) return(pars)
+  pars0 <- pars
+  lp <- length(pars)
+  p <- vector(mode = "character", length = lp)
+  if(is.character(facet)) {
+    facet <- cvec_cs(facet)
+  }
+  if(is.character(group)) {
+    checkvar(group, pars0)
+    p[1] <- group
+    pars <- pars[!pars==group]
+  } else {
+    p[1] <- pars[1]
+    pars <- pars[-1]
+  }
+  if(lp==1) return(p)
+  if(is.character(facet)) {
+    checkvar(facet[1], pars0) 
+    p[2] <- facet[1] 
+    pars <- pars[!pars==facet[1]]
+  } else {
+    p[2] <- pars[1]
+    pars <- pars[-1]
+  }
+  if(lp==2) return(p)
+  if(length(facet) > 1) {
+    checkvar(facet[2], pars0)
+    p[3] <- facet[2]
+  } else {
+    p[3] <- pars[length(pars)]  
+  }
+  if(any(duplicated(p))) {
+    warn("duplicated grouping or faceting variables.")  
+  }
+  p
+}
+
+
 #' Plot sensitivity analysis results
 #' 
 #' @param data output from [sens_each()] or 
@@ -28,7 +86,7 @@ sens_color_n <- function(data, group) {
 #' @param dv_name dependent variable names to plot; can be a comma-separated 
 #' string; if `NULL`, then the unique values of `dv_name` in `data` are used.
 #' @param p_name parameter names to plot; can be a comma-separates string. 
-#' @param logy if `TRUE`, y-axis is transformed to log scale
+#' @param logy if `TRUE`, y-axis is transformed to log scale.
 #' @param ncol passed to [ggplot2::facet_wrap()].
 #' @param lwd passed to [ggplot2::geom_line()].
 #' @param digits used to format numbers on the strips.
@@ -37,6 +95,10 @@ sens_color_n <- function(data, group) {
 #' @param grid if `TRUE`, plots from the `sens_each` method
 #' will be arranged on a page with [patchwork::wrap_plots()]; see the `ncol`
 #' argument.
+#' @param palette a discrete color scale; like what you get from calling
+#' [ggplot2::scale_color_discrete()]. For `sens_each`, this is only applied
+#' when `grid = TRUE`; it is ignored for all other layouts, which use a
+#' continuous viridis color scale.
 #' 
 #' @return 
 #' A `ggplot` object when one `dv_name` is specified or a list of `ggplot` 
@@ -47,9 +109,33 @@ sens_color_n <- function(data, group) {
 #' 
 #' dose <- mrgsolve::ev(amt = 100)
 #' 
-#' out <- sens_run(mod, sargs = list(events = dose),  par = "CL,VC") 
+#' out <- sens_run(
+#'   mod, 
+#'   sargs = list(events = dose),  
+#'   par = "CL,VC"
+#' ) 
 #' 
-#' sens_plot(out, dv_name = "CP")
+#' sens_plot(out, "CP")
+#' 
+#' out <- sens_run(
+#'   mod, 
+#'   sargs = list(events = dose), 
+#'   par = "CL,VC", 
+#'   vary  = "grid", 
+#'   .n = 3
+#' )
+#' 
+#' sens_plot(out, "CP")
+#' 
+#' sens_plot(out, "CP", group = "VC")
+#' 
+#' if(requireNamespace("ggsci")) {
+#' 
+#'   color <- ggsci::scale_color_atlassian()
+#'   
+#'   sens_plot(out, "CP", palette = color)
+#' 
+#' }
 #' 
 #' @export
 sens_plot <- function(data,...) UseMethod("sens_plot")
@@ -86,7 +172,9 @@ sens_plot.sens_each <- function(data, dv_name = NULL, p_name = NULL,
                                 xlab = "time", ylab = NULL,
                                 layout = c("default", "facet_grid", 
                                            "facet_wrap", "list"),
-                                grid = FALSE, ...) {
+                                grid = FALSE, 
+                                palette = NULL, 
+                                ...) {
   
   layout <- match.arg(layout)
   
@@ -219,9 +307,16 @@ sens_plot.sens_each <- function(data, dv_name = NULL, p_name = NULL,
   }
   
   # Grid
+  
   sp <- split(data, data[["p_name"]])
   
   plots <- lapply(sp, function(chunk) {
+    
+    if(is.null(palette)) {
+      ncolor <- length(unique(chunk[["p_value"]]))
+      palette <- pick_palette(ncolor, chunk[["p_name"]][1])
+    }
+    
     chunk[["p_value"]] <- signif(chunk[["p_value"]], digits)
     chunk[["p_value"]] <- factor(chunk[["p_value"]])
     
@@ -232,7 +327,7 @@ sens_plot.sens_each <- function(data, dv_name = NULL, p_name = NULL,
       theme_bw() + xlab(xlab) + ylab(ylab) + 
       facet_wrap(facets = "p_name", scales = "free_y", ncol = ncol) + 
       theme(legend.position = "top") + 
-      scale_color_discrete(name = "")
+      palette + labs(color = chunk[["p_name"]][1])
     if(isTRUE(logy)) {
       p <- p + scale_y_log10()  
     }
@@ -261,6 +356,12 @@ sens_plot_list <- function(dv_name, ylab, args) {
   return(out)
 }
 
+#' @param group sensitivity variable for within-panel grouping; defaults to the 
+#' first sensitivity variable.
+#' @param facet sensitivity variable for faceting when 3 sensitivity variables
+#' are being plotted; the `facet` variable will run left to right and the other
+#' variable will run up and down; this argument is ignored / not needed when 
+#' there are fewer than 3 sensitivity variables.
 #' @rdname sens_plot
 #' @export
 sens_plot.sens_grid <- function(data, 
@@ -272,6 +373,9 @@ sens_plot.sens_grid <- function(data,
                                 plot_ref = TRUE, 
                                 xlab = "time", 
                                 ylab = dv_name,
+                                group = NULL,
+                                facet = NULL,
+                                palette = NULL,
                                 ...) { #nocov start
   
   if(is.null(dv_name)) {
@@ -317,9 +421,14 @@ sens_plot.sens_grid <- function(data,
       )
     )  
   }
-  
   data <- select_sens(data, dv_name = dv_name)
   data <- sens_names_to_factor(data)
+  pars <- sens_grid_plot_vars(pars, group, facet)
+
+  if(is.null(palette)) {
+    ncolor <- length(unique(data[[pars[1]]]))
+    palette <- pick_palette(ncolor, name = pars[1]) 
+  }
   group <- sym(pars[1])
   tcol <- "time"
   if(exists("TIME", data)) tcol <- "TIME"
@@ -337,9 +446,10 @@ sens_plot.sens_grid <- function(data,
     data <- sens_factor(data, pars[3], digits = digits)
   }
   p <- ggplot(data = data, aes(!!x, !!y, group=!!group, col=factor(!!group)))  
-  p <- p + geom_line(lwd=lwd) + scale_color_discrete(name = pars[1])
+  p <- p + geom_line(lwd=lwd) 
   p <- p + theme_bw() + theme(legend.position = "top")
   p <- p + xlab(xlab) + ylab(ylab)
+  p <- p + palette + labs(color = pars[1])
   if(npar==2) p <- p + facet_wrap(formula, ncol = ncol)
   if(npar==3) p <- p + facet_grid(formula)
   if(isTRUE(logy)) p <- p + scale_y_log10()